Mulig å oppdatere en aktivitet fra nytt endepunkt + litt opprydding
